move RPBDWM to test folder from demo
